examples/requirements.txt
openai==1.65.5
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
